python_consumer_producer_pair/requirements.txt
pandas
matplotlib
kafka-python